move : #43 SearchRequest 클래스 PostRequestDTO로 이동
